Add basic API for controlling volume plugins
This is not at all integrated with the rest of Libpod, but has
most of what we need to talk to a volume plugin; the remainder is
all around discovering plugins.

There's some cleanup that can be done in how we handle responses;
a lot can probably be refactored into a shared function, and I
think we probably ought to be checking the response ContentType
to verify the API version is correct (probably not necessary, I
don't think we're likely to see a bump of the API version any
time soon).

Testing this in the repo is going to prove exceedingly difficult,
given that we need to set up an external HTTP server to test
against. Given potentially complexity, I'll do it in a separate
commit.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>