Fix race condition in kill test leading to hang
When you open a FIFO for reading, but there's no writer, you hang.
This is just one of those obscure UNIXisms we all know but just
forget all too often.

My last PR was guilty of introducing such a condition; I caught
it by accident while testing other stuff. In short, the signal
container was doing 'echo DONE' as its last step, and we (BATS)
were reading the FIFO to check for it; but if the container
exited before we opened the FIFO for read, the open would hang.
This is not a hang that we can catch in the test: it would hang
the entire job forever.

Solution: there's no need for the DONE. Just remove it and
the check. (For future reference, another option would be
to 'exec 5<$fifo' immediately after spawning the log reader,
and using 'read -u 5' instead of 'read <$fifo', but that's
unwieldy and hard to maintain.)

Signed-off-by: Ed Santiago <santiago@redhat.com>