Move install.md to podman.io, leave link page
The installation instructions for podman were on this site and a
second copy on podman.io.  I've created https://github.com/containers/podman.io/pull/193
which has a merging of the installation instructions from both sites.

Since the one on podman.io was created a few months ago, we've had a number
of divergences.  Maintaining only one copy should assist with that problem
plus make it easier for people updating them.

This PR should NOT be merged before the one in podman.io.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>